added hook useCarousel
