+++ src/app/events/page.tsx
@@ -6,28 +6,24 @@ import { useSession } from "next-auth/react";
 import { EventFilters } from "@/components/widgets/EventFilters";
 import { CreateAndInvite } from "@/components/widgets/CreateAndInvite";
 import { Mode } from "@/components/widgets/CreateAndInvite/CreateAndInvite";
+import { useEventList } from "@/components/widgets/EventList/hooks";
 
 import { Title } from "@/components/shared/Title";
-import { Slider } from "@/components/shared/Slider";
 import { Carousel } from "@/components/shared/Carousel";
+import { Slide } from "@/components/shared/Slide";
+
+import { IEvent } from "@/database/models/Event";
 
 import classes from "./page.module.css";
-import { useSlider } from "@/components/shared/Slider/hook/useSlider";
-import { IEvent } from "@/database/models/Event";
-import { Slide } from "@/components/shared/Slider/Slide";
 
 const EventListPage: NextPage = () => {
     const { status } = useSession();
-    const { hideSliderHandler, hideSlider, currentIndex, filterEventsImage, slideWidth, nextSlide } = useSlider();
-
-    const arr = filterEventsImage.length;
+    const { filterEventsImage } = useEventList({});
 
     return (
         <div className={classes.eventListPage}>
             <Title tag={"h1"} title="Присоединяйтесь к приключениям" className={classes.title} />
-            {/* <Slider /> */}
-            {/* <div style={{ width: "100%", marginLeft: "auto", marginRight: "auto", marginTop: 64 }}> */}
-            <Carousel show={3} arr={arr}>
+            <Carousel>
                 {filterEventsImage.map((slide: IEvent) => (
                     <Slide
                         key={slide._id}
@@ -40,7 +36,6 @@ const EventListPage: NextPage = () => {
                     />
                 ))}
             </Carousel>
-            {/* </div> */}
 
             <EventFilters />
             <CreateAndInvite mode={Mode.EVENT} status={status} />
