+++ src/components/shared/Carousel/Carousel.tsx
@@ -1,45 +1,15 @@
-import { FC, useEffect, useState } from "react";
+import { FC } from "react";
 
 import { Title } from "@/components/shared/Title";
 import { Button } from "@/components/shared/Button";
 import { ButtonNext } from "@/components/shared/ButtonNext";
 
+import { ICarousel, useCarousel } from "./hooks/useCarousel";
+
 import classes from "./Carousel.module.css";
 
-interface ICarousel {
-    children: React.ReactNode[];
-}
-
 export const Carousel: FC<ICarousel> = ({ children }) => {
-    const [width, setWidth] = useState<number | null>(null);
-    const [hideSlider, setHideSlider] = useState(false);
-    const [currentIndex, setCurrentIndex] = useState<number>(0);
-    const [length, setLength] = useState<number>(children.length);
-
-    const hideSliderHandler = () => {
-        setHideSlider(true);
-    };
-
-    const next = () => {
-        setCurrentIndex((prevIndex) => (prevIndex + 1) % length);
-    };
-
-    useEffect(() => {
-        setLength(children.length);
-    }, [children]);
-
-    useEffect(() => {
-        const updateWidth = () => {
-            const windowInnerWidth = window.innerWidth;
-            const containerGrid = 1180;
-            setWidth((windowInnerWidth - containerGrid) / 2 + containerGrid);
-        };
-
-        updateWidth();
-        window.addEventListener("resize", updateWidth);
-
-        return () => window.removeEventListener("resize", updateWidth);
-    }, []);
+    const { width, hideSlider, currentIndex, next, hideSliderHandler, length } = useCarousel({ children });
 
     return (
         <>
