+++ src/components/shared/Slide/Slide.module.css
