+++ src/components/shared/Slide/Slide.tsx
