+++ src/components/shared/Slide/index.ts
