+++ src/components/widgets/EventList/hooks/useEventList.ts
@@ -1,6 +1,7 @@
 import { GET_EVENTS } from "@/app/api/graphql/queries/events";
 import { useQuery } from "@apollo/client";
 import { GetEventsData, useEventListProps } from "@/components/widgets/EventList/types/EventList";
+import { IEvent } from "@/database/models/Event";
 
 export const useEventList = ({ limit, offset, sort }: useEventListProps) => {
     const { loading, error, data, refetch } = useQuery<GetEventsData>(GET_EVENTS, {
@@ -13,5 +14,9 @@ export const useEventList = ({ limit, offset, sort }: useEventListProps) => {
 
     const events = data?.events;
 
-    return { loading, error, events, refetch };
+    const filterEventsImage: IEvent[] = events
+        ? events.filter((event: IEvent) => event.image !== undefined && event.image !== "")
+        : [];
+
+    return { loading, error, events, refetch, filterEventsImage };
 };
